Improve branch cleanup test clarity and update Jest configuration
